+++ docs-web/.vitepress/theme/Layout.vue
@@ -21,7 +21,7 @@ const { Layout } = DefaultTheme
       <PlatformCompat />
     </template>
     <template #nav-bar-title-after>
-      <EvTag size="small" style="margin-left: 10px;">v0.12.0</EvTag>
+      <EvTag size="small" style="margin-left: 10px;">v0.13.0</EvTag>
     </template>
     <template #nav-bar-content-after>
       <span class="home-nav-actions">